cleanup code add new header file with global enumaration constants

+++ bld/wipfc/cpp/uniutil.cpp
@@ -93,7 +93,7 @@ std::size_t wtomb_cstring( char *dst_mbc, const wchar_t *src_wc, std::size_t len
 
     while( len > 0 && *src_wc != L'\0' ) {
         bytes = wtomb_char( mbc, *src_wc );
-        if( bytes == -1 || bytes > len )
+        if( bytes == -1 || (unsigned)bytes > len )
             return( static_cast<std::size_t>( -1 ) );
         std::memcpy( dst_mbc, mbc, bytes );
         dst_mbc += bytes;
